Separate the lightcurve database into subclasses for the two microlensing data types
