+++ microlensing_label_per_time_step_database.py
@@ -1,0 +1,7 @@
+"""Code for representing a dataset of lightcurves for binary classification with a single label per time step."""
+from lightcurve_database import LightcurveDatabase
+
+
+class MicrolensingLabelPerTimeStepDatabase(LightcurveDatabase):
+    """A representing a dataset of lightcurves for binary classification with a single label per time step."""
+
